Move af_common module into neuralplexer package
