+++ tests/test_template_stack.py
@@ -32,7 +32,7 @@ from neuralplexer.util.frame import (
     get_frame_matrix,
 )
 from neuralplexer.util.pdb3d import get_lddt_bs
-from af_common.protein import Protein, from_prediction, to_pdb
+from neuralplexer.af_common.protein import Protein, from_prediction, to_pdb
 from rdkit.Chem import AllChem
 from rdkit.Geometry import Point3D
 from rdkit.Chem.rdmolops import CombineMols
